Reorg
Making things more modular for easy of maintenance and writing tests.
